refactor: 사용하지 않는 파일 제거
