+++ apps/web/src/app/[locale]/shop/SearchOption/index.ts
@@ -1,5 +1,4 @@
 import OrderTypeSelect from './OrderTypeSelect';
-import PersonaType from './PersonaType';
 import SortDirectionSelect from './SortDirectionSelect';
 
-export { OrderTypeSelect, PersonaType, SortDirectionSelect };
+export { OrderTypeSelect, SortDirectionSelect };
